Switch expt setup to use case files
Preliminarily updates experiment setup to use case files instead of a
bunch of command line switches to define experiments. In addition to
the settings previously determined by experiment setup, it also allows
for the addition of an arbitrary list of variables that will be written
as a config file to be ingested at runtime.

The command line syntax is substantially updated and streamlined. The
positional argument for mode is removed. Five named switches remain.
Three that operate the same as before
- --pslot: optional argument giving the experiment name
- --comrot: optional argument giving the top of the COM directory
- --expdir: optional argument giving the top of the expt directory
And two that are new:
- --case_file: a mandatory argument providing the YAML file containng
               the settings for the experiment
- --user_file: an optional argument providing a YAML with user-specific
               settings (comrot and expdir) rather than using the
               command line

Settings have the following precedence:
command line > case_file > user_file > (defaults)

So that settings on the command line override those in the case_file,
etc. While user_file and case_file are designed to hold different, non-
overlapping settings, the processing treats them almost the same, so
settings in case_file would override those in the user_file, if present.
(case_file idate and edate are converted into datetime objects, but the
same is not true of user_file.)

The case_file must contain three settings: mode (cycled or forecast-only),
idate, and edate. Additionally, all of the settings previously available
through the command line are available (app, resdet, etc). Finally, a
'settings' category can be added. And variable-value pairs listed within
settings will be converted into a new config.case file in the experiment
directory that in the future will be used to provide overrides to experi-
ment settings. A temporary test case_file is added here, but more will be
added when the system is complete.

The user_file is designed to hold the user-specific settings that would
not be approriate in a case file (currently rotdir and expdir).

After all settings have been loaded, they are validated to ensure all
mandatory variables (mode, idate, edate) are set, certain variables have
valid settings, etc. This is likely more properly done using schema; that
may be done at a future time but the default python installation on Hera
does not currently include that library.

Refs: #710, #711