Update config.base to load config.case, APP changes
Updates config.base to source the config.case file if present. Since
this is done first and expdir is not yet defined, the location is
derived from the script location. As part of this, the script is
changed to bash.

Also reverts the previous change that had setup_expt.py use the
--aerosols switch to set DO_AERO, instead new APPs are added for
aerosols.

Finally, settings are updated to accept overrides so settings from
config.case will be properly applied.

Refs: #710, #711