Replace old post script with initial post refactor
Refactors the post job to only run the offline post portion.
Interpolation of grib files will be moved to a new product
generation job using less resources. So will WAFS/GOES post,
though those will reuse the base python script.

New script is primarily python-based. The shell exscript is
responsible for providing the python script with the appropriate
YAML settings file for the cdump and forecast hour. The YAML
file, located in `parm/post`, is responsible for providing all
of the needed settings (many via environment variables).

This is still a work in progress. Only offline post for gfs
forecast hours is confirmed to work. Support for other cdumps,
as well as WAFS and GOES, will be implemented later.

Refs: #270