Add sflux to post output
Adds the creation of surface flux files to the post job. Working
directories have 'atm' or 'sflux' added to the path so each has
its own for debugging purposes.

Copying of files into the working directory has been changed to
symlinks to prevent the increase in resources by multiple copies
of the model output.

Refs: #270